// Soe.Angular/TypeScript/Common/Models/CustomerIODTO.ts
import { IContactAdressIODTO, IContactEComIODTO, ICustomerIODTO } from "../../Scripts/TypeLite.Net4";
import { TermGroup_IOType, TermGroup_IOStatus, TermGroup_IOSource, TermGroup_IOImportHeadType } from "../../Util/CommonEnumerations";

export class CustomerIODTO implements ICustomerIODTO {
    accountsReceivableAccountInternal1: string;
    accountsReceivableAccountInternal2: string;
    accountsReceivableAccountInternal3: string;
    accountsReceivableAccountInternal4: string;
    accountsReceivableAccountInternal5: string;
    accountsReceivableAccountNr: string;
    accountsReceivableAccountSieDim1: string;
    accountsReceivableAccountSieDim6: string;
    actorCompanyId: number;
    agreementTemplate: number;
    batchId: string;
    billingAddresses: IContactAdressIODTO[];
    billingAddress: string;
    billingCoAddress: string;
    billingCountry: string;
    billingPostalAddress: string;
    billingPostalCode: string;
    billingTemplate: number;
    blockInvoice: boolean;
    blockNote: string;
    blockOrder: boolean;
    boardHQAddress: string;
    boardHQCountry: string;
    categoryCode1: string;
    categoryCode2: string;
    categoryCode3: string;
    categoryCode4: string;
    categoryCode5: string;
    categoryIds: number[];
    contactEmail: string;
    contactFirstName: string;
    contactLastName: string;
    country: string;
    created: Date;
    createdBy: string;
    creditLimit: number;
    currency: string;
    customerId: number;
    customerIOId: number;
    customerNr: string;
    customerState: number;
    defaultPriceListType: string;
    defaultPriceListTypeId: number;
    defaultWholeseller: string;
    deliveryAddress: string;
    deliveryCoAddress: string;
    deliveryCondition: string;
    deliveryCountry: string;
    deliveryMethod: string;
    deliveryPostalAddress: string;
    deliveryPostalCode: string;
    disableInvoiceFee: boolean;
    discountMerchandise: number;
    discountService: number;
    distributionAddress: string;
    distributionCoAddress: string;
    distributionCountry: string;
    distributionPostalAddress: string;
    distributionPostalCode: string;
    email1: string;
    email2: string;
    errorMessage: string;
    externalNbrs: string[];
    fax: string;
    finvoiceAddress: string;
    finvoiceOperator: string;
    gln: string;
    gLNNbrs: IContactEComIODTO[];
    gracePeriodDays: number;
    import: boolean;
    importHeadType: TermGroup_IOImportHeadType;
    importInvoiceDetailed: boolean;
    invoiceDeliveryEmail: string;
    invoiceDeliveryType: number;
    invoiceLabel: string;
    invoiceReference: string;
    isCashCustomer: boolean;
    isFinvoiceCustomer: boolean;
    isModified: boolean;
    isPrivatePerson: boolean;
    isSelected: boolean;
    language: string;
    manualAccounting: boolean;
    modified: Date;
    modifiedBy: string;
    name: string;
    note: string;
    offerTemplate: number;
    orderTemplate: number;
    orgNr: string;
    paymentCondition: string;
    phoneHome: string;
    phoneJob: string;
    phoneMobile: string;
    productPrices: any[];
    salesAccountInternal1: string;
    salesAccountInternal2: string;
    salesAccountInternal3: string;
    salesAccountInternal4: string;
    salesAccountInternal5: string;
    salesAccountNr: string;
    salesAccountSieDim1: string;
    salesAccountSieDim6: string;
    showNote: boolean;
    source: TermGroup_IOSource;
    state: number;
    status: TermGroup_IOStatus;
    statusName: string;
    supplierNr: number;
    sysLanguageId: number;
    type: TermGroup_IOType;
    vATAccountNr: string;
    vATCodeNr: string;
    vatNr: string;
    vatType: number;
    vatTypeName: string;
    webpage: string;
    visitingAddress: string;
    visitingCoAddress: string;
    visitingCountry: string;
    visitingPostalAddress: string;
    visitingPostalCode: string;
}